+++ appObjects/FlatCAMGeometry.py
@@ -657,6 +657,7 @@ class GeometryObject(FlatCAMObj, Geometry):
         self.ui.generate_cnc_button.clicked.connect(self.on_generatecnc_button_click)
         self.ui.paint_tool_button.clicked.connect(lambda: self.app.paint_tool.run(toggle=False))
         self.ui.generate_ncc_button.clicked.connect(lambda: self.app.ncclear_tool.run(toggle=False))
+        self.ui.milling_button.clicked.connect(self.on_milling_button_clicked)
 
         # Postprocessor change
         self.ui.pp_geometry_name_cb.activated.connect(self.on_pp_changed)
@@ -697,6 +698,12 @@ class GeometryObject(FlatCAMObj, Geometry):
         self.ui.treeWidget.setSizePolicy(QtWidgets.QSizePolicy.Ignored, QtWidgets.QSizePolicy.MinimumExpanding)
         # make sure that the FCTree widget columns are resized to content
         self.ui.treeWidget.resize_sig.emit()
+
+    def on_milling_button_clicked(self):
+        self.app.milling_tool.run(toggle=False)
+        self.app.milling_tool.ui.target_radio.set_value('geo')
+        current_obj = self.app.collection.get_active()
+        self.app.milling_tool.ui.object_combo.set_value(current_obj.options['name'])
 
     def on_rebuild_ui(self):
         # read the table tools uid