+++ CHANGELOG.md
@@ -17,6 +17,7 @@ CHANGELOG for FlatCAM beta
 - minimized the number of tools present in the Tools Toolbar to the bare minimum
 - fixed an error in the Follow Tool
 - started to work in Milling Tool - designing the Tool UI
+- Milling Tool - finished the UI and also the button handlers in the Geometry and Excellon objects Properties UI
 
 12.11.2020